move custom helper functions to its test file
