+++ tests/Datasets/Attributes/RulesDataset.php
@@ -85,402 +85,417 @@ use Spatie\LaravelData\Attributes\Validation\Unique;
 use Spatie\LaravelData\Attributes\Validation\Url;
 use Spatie\LaravelData\Attributes\Validation\Uuid;
 use Spatie\LaravelData\Exceptions\CannotBuildValidationRule;
+use Spatie\LaravelData\Support\Validation\ValidationRule;
+
+function fixture(
+    ValidationRule $attribute,
+    object|string|array $expected,
+    object|string|null $expectCreatedAttribute = null,
+    string $exception = null
+) {
+    return [
+        'attribute' => $attribute,
+        'expected' => $expected,
+        'expectedCreatedAttribute' => $expectCreatedAttribute ?? $attribute,
+        'exception' => $exception,
+    ];
+}
 
 dataset('attributes', function () {
-    yield rulesFixture(
+    yield fixture(
         attribute: new Accepted(),
         expected: 'accepted',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new ActiveUrl(),
         expected: 'active_url',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Alpha(),
         expected: 'alpha',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new AlphaDash(),
         expected: 'alpha_dash',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new AlphaNumeric(),
         expected: 'alpha_num',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Bail(),
         expected: 'bail',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new BooleanType(),
         expected: 'boolean',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Confirmed(),
         expected: 'confirmed',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Date(),
         expected: 'date',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new DateFormat('Y-m-d'),
         expected: 'date_format:Y-m-d',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Different('field'),
         expected: 'different:field',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Digits(10),
         expected: 'digits:10',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new DigitsBetween(5, 10),
         expected: 'digits_between:5,10',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Enum('enum_class'),
         expected: new EnumRule('enum_class'),
         expectCreatedAttribute: new Enum(new EnumRule('enum_class'))
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new ExcludeIf('field', true),
         expected: 'exclude_if:field,true',
         expectCreatedAttribute: new ExcludeIf('field', true)
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new ExcludeUnless('field', 42),
         expected: 'exclude_unless:field,42',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new ExcludeWithout('field'),
         expected: 'exclude_without:field',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new File(),
         expected: 'file',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Filled(),
         expected: 'filled',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new GreaterThan('field'),
         expected: 'gt:field',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new GreaterThanOrEqualTo('field'),
         expected: 'gte:field',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Image(),
         expected: 'image',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new InArray('field'),
         expected: 'in_array:field',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new IntegerType(),
         expected: 'integer',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new IP(),
         expected: 'ip',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new IPv4(),
         expected: 'ipv4',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new IPv6(),
         expected: 'ipv6',
     );
 
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Json(),
         expected: 'json',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new LessThan('field'),
         expected: 'lt:field',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new LessThanOrEqualTo('field'),
         expected: 'lte:field',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Max(10),
         expected: 'max:10',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Min(10),
         expected: 'min:10',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new MultipleOf(10),
         expected: 'multiple_of:10',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new NotRegex('/^.+$/i'),
         expected: 'not_regex:/^.+$/i',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Nullable(),
         expected: 'nullable',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Numeric(),
         expected: 'numeric',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Present(),
         expected: 'present',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Prohibited(),
         expected: 'prohibited',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Regex('/^.+$/i'),
         expected: 'regex:/^.+$/i',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Required(),
         expected: 'required',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Same('field'),
         expected: 'same:field',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Size(10),
         expected: 'size:10',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new StringType(),
         expected: 'string',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Timezone(),
         expected: 'timezone',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Url(),
         expected: 'url',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Uuid(),
         expected: 'uuid',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Sometimes(),
         expected: 'sometimes',
     );
 
     // Accepted if attributes
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new AcceptedIf('value', 'string'),
         expected: 'accepted_if:value,string',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new AcceptedIf('value', true),
         expected: 'accepted_if:value,true',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new AcceptedIf('value', 42),
         expected: 'accepted_if:value,42',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new AcceptedIf('value', 3.14),
         expected: 'accepted_if:value,3.14',
     );
 
     // After Attributes
 
-     yield rulesFixture(
+    yield fixture(
         attribute: new After('some_field'),
         expected: 'after:some_field',
     );
 
-        yield rulesFixture(
+    yield fixture(
         attribute: new After('2020-05-15 00:00:00'),
         expected: 'after:2020-05-15 00:00:00',
     );
 
     // After or equal attributes
 
-         yield rulesFixture(
+    yield fixture(
         attribute: new AfterOrEqual('some_field'),
         expected: 'after_or_equal:some_field',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new AfterOrEqual('2020-05-15 00:00:00'),
         expected: 'after_or_equal:2020-05-15 00:00:00',
     );
 
     // Array type attributes
 
-      yield rulesFixture(
+    yield fixture(
         attribute: new ArrayType(),
         expected: 'array',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new ArrayType(['a', 'b', 'c']),
         expected: 'array:a,b,c',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new ArrayType('a', 'b', 'c'),
         expected: 'array:a,b,c',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new ArrayType('a', ['b', 'c']),
         expected: 'array:a,b,c',
     );
 
     // Before attributes
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Before('some_field'),
         expected: 'before:some_field',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Before('2020-05-15 00:00:00'),
         expected: 'before:2020-05-15 00:00:00',
     );
 
     // Before or equal attributes
 
-     yield rulesFixture(
+    yield fixture(
         attribute: new BeforeOrEqual('some_field'),
         expected: 'before_or_equal:some_field',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new BeforeOrEqual('2020-05-15 00:00:00'),
         expected: 'before_or_equal:2020-05-15 00:00:00',
     );
 
     // Between attributes
 
-     yield rulesFixture(
+    yield fixture(
         attribute: new Between(-42, 42),
         expected: 'between:-42,42',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Between(-3.14, 3.14),
         expected: 'between:-3.14,3.14',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Between(-3.14, 42),
         expected: 'between:-3.14,42',
     );
 
     // Current password attributes
 
-     yield rulesFixture(
+    yield fixture(
         attribute: new CurrentPassword(),
         expected: 'current_password',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new CurrentPassword('api'),
         expected: 'current_password:api',
     );
 
     // Date equals attributes
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new DateEquals('tomorrow'),
         expected: 'date_equals:tomorrow',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new DateEquals('2020-05-15 00:00:00'),
         expected: 'date_equals:2020-05-15 00:00:00',
     );
 
     // Dimenstions attributes
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Dimensions(minWidth: 15, minHeight: 10, maxWidth: 150, maxHeight: 100, ratio: 1),
         expected: 'dimensions:min_width=15,min_height=10,max_width=150,max_height=100,ratio=1',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Dimensions(maxWidth: 150, maxHeight: 100),
         expected: 'dimensions:max_width=150,max_height=100',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Dimensions(ratio: 1.5),
         expected: 'dimensions:ratio=1.5',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Dimensions(ratio: '3/4'),
         expected: 'dimensions:ratio=3/4',
     );
 
-    //        yield rulesFixture(
+    //        yield fixture(
     //            attribute: new Dimensions(),
     //            expected: '',
     //            exception: CannotBuildValidationRule::class,
@@ -488,22 +503,22 @@ dataset('attributes', function () {
 
     // Distinct attributes
 
-     yield rulesFixture(
+    yield fixture(
         attribute: new Distinct(),
         expected: 'distinct',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Distinct(Distinct::Strict),
         expected: 'distinct:strict',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Distinct(Distinct::IgnoreCase),
         expected: 'distinct:ignore_case',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Distinct('fake'),
         expected: '',
         exception: CannotBuildValidationRule::class
@@ -511,33 +526,33 @@ dataset('attributes', function () {
 
     // Email attributes
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Email(),
         expected: 'email:rfc',
         expectCreatedAttribute: new Email(Email::RfcValidation),
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Email(Email::RfcValidation),
         expected: 'email:rfc',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Email([Email::RfcValidation, Email::NoRfcWarningsValidation]),
         expected: 'email:rfc,strict',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Email([Email::RfcValidation, Email::NoRfcWarningsValidation, Email::DnsCheckValidation, Email::SpoofCheckValidation, Email::FilterEmailValidation]),
         expected: 'email:rfc,strict,dns,spoof,filter',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Email(Email::RfcValidation, Email::NoRfcWarningsValidation),
         expected: 'email:rfc,strict',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Email(['fake']),
         expected: '',
         exception: CannotBuildValidationRule::class,
@@ -545,36 +560,36 @@ dataset('attributes', function () {
 
     // Ends with attributes
 
-      yield rulesFixture(
+    yield fixture(
         attribute: new EndsWith('x'),
         expected: 'ends_with:x',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new EndsWith(['x', 'y']),
         expected: 'ends_with:x,y',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new EndsWith('x', 'y'),
         expected: 'ends_with:x,y',
     );
 
     // Exists attributes
 
-       yield rulesFixture(
+    yield fixture(
         attribute: new Exists('users'),
         expected: new BaseExists('users'),
         expectCreatedAttribute: new Exists(rule: new BaseExists('users'))
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Exists('users', 'email'),
         expected: new BaseExists('users', 'email'),
         expectCreatedAttribute: new Exists(rule: new BaseExists('users', 'email'))
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Exists('users', 'email', connection: 'tenant'),
         expected: new BaseExists('tenant.users', 'email'),
         expectCreatedAttribute: new Exists(rule: new BaseExists('tenant.users', 'email'))
@@ -582,7 +597,7 @@ dataset('attributes', function () {
 
     $closure = fn (Builder $builder) => $builder;
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Exists('users', 'email', where: $closure),
         expected: (new BaseExists('users', 'email'))->where($closure),
         expectCreatedAttribute: new Exists(rule: (new BaseExists('users', 'email'))->where($closure))
@@ -590,19 +605,19 @@ dataset('attributes', function () {
 
     // In attributes
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new In('key'),
         expected: new BaseIn(['key']),
         expectCreatedAttribute: new In(new BaseIn(['key']))
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new In(['key', 'other']),
         expected: new BaseIn(['key', 'other']),
         expectCreatedAttribute: new In(new BaseIn(['key', 'other']))
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new In('key', 'other'),
         expected: new BaseIn(['key', 'other']),
         expectCreatedAttribute: new In(new BaseIn(['key', 'other']))
@@ -610,53 +625,53 @@ dataset('attributes', function () {
 
     // Mime attributes
 
-     yield rulesFixture(
+    yield fixture(
         attribute: new MimeTypes('video/quicktime'),
         expected: 'mimestypes:video/quicktime',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new MimeTypes(['video/quicktime', 'video/avi']),
         expected: 'mimestypes:video/quicktime,video/avi',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new MimeTypes('video/quicktime', 'video/avi'),
         expected: 'mimestypes:video/quicktime,video/avi',
     );
 
     // Mime types attributes
 
-     yield rulesFixture(
+    yield fixture(
         attribute: new Mimes('jpg'),
         expected: 'mimes:jpg',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Mimes(['jpg', 'png']),
         expected: 'mimes:jpg,png',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Mimes('jpg', 'png'),
         expected: 'mimes:jpg,png',
     );
 
     // Not in attributes
 
-     yield rulesFixture(
+    yield fixture(
         attribute: new NotIn('key'),
         expected: new BaseNotIn(['key']),
         expectCreatedAttribute: new NotIn(new BaseNotIn(['key']))
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new NotIn(['key', 'other']),
         expected: new BaseNotIn(['key', 'other']),
         expectCreatedAttribute: new NotIn(new BaseNotIn(['key', 'other']))
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new NotIn('key', 'other'),
         expected: new BaseNotIn(['key', 'other']),
         expectCreatedAttribute: new NotIn(new BaseNotIn(['key', 'other']))
@@ -664,19 +679,19 @@ dataset('attributes', function () {
 
     // Password attributes
 
-     yield rulesFixture(
+    yield fixture(
         attribute: new Password(),
         expected: new BasePassword(12),
         expectCreatedAttribute: new Password(rule: new BasePassword(12)),
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Password(min: 20),
         expected: new BasePassword(20),
         expectCreatedAttribute: new Password(rule: new BasePassword(20)),
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Password(letters: true, mixedCase: true, numbers: true, uncompromised: true, uncompromisedThreshold: 12),
         expected: (new BasePassword(12))->letters()->mixedCase()->numbers()->uncompromised(12),
         expectCreatedAttribute: new Password(rule: (new BasePassword(12))->letters()->mixedCase()->numbers()->uncompromised(12),),
@@ -684,235 +699,235 @@ dataset('attributes', function () {
 
     // Prohibited if attributes
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new ProhibitedIf('field', 'key'),
         expected: 'prohibited_if:field,key',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new ProhibitedIf('field', ['key', 'other']),
         expected: 'prohibited_if:field,key,other',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new ProhibitedIf('field', 'key', 'other'),
         expected: 'prohibited_if:field,key,other',
     );
 
     // Prohibited unless attributes
 
-     yield rulesFixture(
+    yield fixture(
         attribute: new ProhibitedUnless('field', 'key'),
         expected: 'prohibited_unless:field,key',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new ProhibitedUnless('field', ['key', 'other']),
         expected: 'prohibited_unless:field,key,other',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new ProhibitedUnless('field', 'key', 'other'),
         expected: 'prohibited_unless:field,key,other',
     );
 
     // Prohibited attributes
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Prohibits('key'),
         expected: 'prohibits:key',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Prohibits(['key', 'other']),
         expected: 'prohibits:key,other',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Prohibits('key', 'other'),
         expected: 'prohibits:key,other',
     );
 
     // Required if attributes
 
-      yield rulesFixture(
+    yield fixture(
         attribute: new RequiredIf('field'),
         expected: 'required_if:field',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new RequiredIf('field', 'key'),
         expected: 'required_if:field,key',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new RequiredIf('field', ['key', 'other']),
         expected: 'required_if:field,key,other',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new RequiredIf('field', 'key', 'other'),
         expected: 'required_if:field,key,other',
     );
 
     // Required unless attributes
 
-     yield rulesFixture(
+    yield fixture(
         attribute: new RequiredUnless('field', 'key'),
         expected: 'required_unless:field,key',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new RequiredUnless('field', ['key', 'other']),
         expected: 'required_unless:field,key,other',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new RequiredUnless('field', 'key', 'other'),
         expected: 'required_unless:field,key,other',
     );
 
     // Required with attributes
 
-      yield rulesFixture(
+    yield fixture(
         attribute: new RequiredWith('key'),
         expected: 'required_with:key',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new RequiredWith(['key', 'other']),
         expected: 'required_with:key,other',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new RequiredWith('key', 'other'),
         expected: 'required_with:key,other',
     );
 
     // Required with all attributes
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new RequiredWithAll('key'),
         expected: 'required_with_all:key',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new RequiredWithAll(['key', 'other']),
         expected: 'required_with_all:key,other',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new RequiredWithAll('key', 'other'),
         expected: 'required_with_all:key,other',
     );
 
     // Required with all attributes
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new RequiredWithAll('key'),
         expected: 'required_with_all:key',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new RequiredWithAll(['key', 'other']),
         expected: 'required_with_all:key,other',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new RequiredWithAll('key', 'other'),
         expected: 'required_with_all:key,other',
     );
 
     // Required without attributes
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new RequiredWithout('key'),
         expected: 'required_without:key',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new RequiredWithout(['key', 'other']),
         expected: 'required_without:key,other',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new RequiredWithout('key', 'other'),
         expected: 'required_without:key,other',
     );
 
     // Required without all attributes
 
-     yield rulesFixture(
+    yield fixture(
         attribute: new RequiredWithoutAll('key'),
         expected: 'required_without_all:key',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new RequiredWithoutAll(['key', 'other']),
         expected: 'required_without_all:key,other',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new RequiredWithoutAll('key', 'other'),
         expected: 'required_without_all:key,other',
     );
 
     // Starts with attributes
 
-       yield rulesFixture(
+    yield fixture(
         attribute: new StartsWith('x'),
         expected: 'starts_with:x',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new StartsWith(['x', 'y']),
         expected: 'starts_with:x,y',
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new StartsWith('x', 'y'),
         expected: 'starts_with:x,y',
     );
 
     // Unique attributes
 
-       yield rulesFixture(
+    yield fixture(
         attribute: new Unique('users'),
         expected: new BaseUnique('users'),
         expectCreatedAttribute: new Unique(rule: new BaseUnique('users'))
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Unique('users', 'email'),
         expected: new BaseUnique('users', 'email'),
         expectCreatedAttribute: new Unique(rule: new BaseUnique('users', 'email'))
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Unique('users', 'email', connection: 'tenant'),
         expected: new BaseUnique('tenant.users', 'email'),
         expectCreatedAttribute: new Unique(rule: new BaseUnique('tenant.users', 'email'))
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Unique('users', 'email', withoutTrashed: true),
         expected: (new BaseUnique('users', 'email'))->withoutTrashed(),
         expectCreatedAttribute: new Unique(rule: (new BaseUnique('users', 'email'))->withoutTrashed())
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Unique('users', 'email', withoutTrashed: true, deletedAtColumn: 'deleted_when'),
         expected: (new BaseUnique('users', 'email'))->withoutTrashed('deleted_when'),
         expectCreatedAttribute: new Unique(rule: (new BaseUnique('users', 'email'))->withoutTrashed('deleted_when'))
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Unique('users', 'email', ignore: 5),
         expected: (new BaseUnique('users', 'email'))->ignore(5),
         expectCreatedAttribute: new Unique(rule: (new BaseUnique('users', 'email'))->ignore(5))
     );
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Unique('users', 'email', ignore: 5, ignoreColumn: 'uuid'),
         expected: (new BaseUnique('users', 'email'))->ignore(5, 'uuid'),
         expectCreatedAttribute: new Unique(rule: (new BaseUnique('users', 'email'))->ignore(5, 'uuid'))
@@ -920,7 +935,7 @@ dataset('attributes', function () {
 
     $closure = fn (Builder $builder) => $builder;
 
-    yield rulesFixture(
+    yield fixture(
         attribute: new Unique('users', 'email', where: $closure),
         expected: (new BaseUnique('users', 'email'))->where($closure),
         expectCreatedAttribute: new Unique(rule: (new BaseUnique('users', 'email'))->where($closure))
